added some keybindings
added bindings for volume. (up, down and mute toggle)
added binding for play / pause toggle
added binding to open systemsettings gui
added binding to toggle wifi on and off

keep in mind that some of these binds and such might not work for others. as i have a plasma install aswell. as i play games on plasma. but live in qtile.